fix: 6 critical bugs + strengthen autonomous login flow + framework integration
BUG-1: Fix fatal _snapshot -> _save_snapshot AttributeError (3 occurrences)
BUG-2: ExecutorAgent.close() now detects parasitic mode, won't kill external browser
BUG-3: Pipeline _vision_locate coords int->float for Gemma compatibility
BUG-4: Add detach_page() to clean executor state after parasitic task
BUG-5: Login flow unified: auto-login success always queues page + human fallback on failure
BUG-6: Fix JS injection risk in select action via safe argument passing

Strengthened exploration:
- Login detection now has 3-tier fallback: orchestrator -> vision loop -> human intervention
- Orchestrator run_task_on_page uses finally/detach to guarantee clean state

+++ src/webagent/pipeline/pipeline.py
@@ -312,7 +312,7 @@ class ActionPipeline:
 
             if data and data.get("found"):
                 coords = data.get("coordinates", {})
-                llm_x, llm_y = int(coords.get("x", 0)), int(coords.get("y", 0))
+                llm_x, llm_y = float(coords.get("x", 0)), float(coords.get("y", 0))
                 confidence = data.get("confidence", 0)
                 desc = data.get("element_description", "")
